requirements.txt
# MarketHub Django E-Commerce Platform Requirements

# Core Django and Database
Django==5.2.5
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

# Django REST Framework for API
djangorestframework==3.16.1

# Image processing for product photos
Pillow==11.3.0

# Environment configuration management
python-decouple==3.8

# Production deployment dependencies (uncomment for production)
# psycopg2-binary==2.9.7  # PostgreSQL database adapter
# gunicorn==21.2.0        # WSGI HTTP Server
# django-redis==5.3.0     # Redis caching backend
# django-cors-headers==4.2.0  # CORS handling for API
# whitenoise==6.5.0       # Static file serving
